+++ rae_hw/CMakeLists.txt
@@ -14,7 +14,6 @@ ament_auto_find_build_dependencies()
 find_package(ALSA REQUIRED)
 find_package(std_srvs REQUIRED)
 
-
 find_library(GPIOD_LIBRARY NAMES libgpiodcxx.so)
 if(NOT GPIOD_LIBRARY)
   message(FATAL_ERROR "gpiod library not found. Install apt install libgpiod-dev")
@@ -24,15 +23,9 @@ endif()
 find_package(PkgConfig REQUIRED)
 pkg_check_modules(MPG123 REQUIRED libmpg123)
 
-
-
 # Find SNDFILE library
 find_path(SNDFILE_INCLUDE_DIR sndfile.h)
 find_library(SNDFILE_LIBRARY NAMES sndfile)
-
-# Add GStreamer packages
-pkg_check_modules(GST REQUIRED gstreamer-1.0)
-
 
 include_directories(include ${GST_INCLUDE_DIRS} ${MPG123_INCLUDE_DIRS} ${SNDFILE_INCLUDE_DIR})
 
@@ -44,7 +37,6 @@ ament_auto_add_library(
   src/peripherals/battery.cpp
   src/peripherals/lcd.cpp
   src/peripherals/led.cpp
-  src/peripherals/mic.cpp
 )
 
 ament_target_dependencies(${PROJECT_NAME} ${DEPENDENCIES} ALSA std_srvs)
@@ -62,7 +54,6 @@ target_compile_definitions(${PROJECT_NAME} PUBLIC "PLUGINLIB__DISABLE_BOOST_FUNC
 rclcpp_components_register_nodes(${PROJECT_NAME} "${PROJECT_NAME}::BatteryNode")
 rclcpp_components_register_nodes(${PROJECT_NAME} "${PROJECT_NAME}::LCDNode")
 rclcpp_components_register_nodes(${PROJECT_NAME} "${PROJECT_NAME}::LEDNode")
-rclcpp_components_register_nodes(${PROJECT_NAME} "${PROJECT_NAME}::MicNode")
 rclcpp_components_register_nodes(${PROJECT_NAME} "${PROJECT_NAME}::SpeakersNode")
 
 pluginlib_export_plugin_description_file(
@@ -83,10 +74,6 @@ target_link_libraries(mic_node ${PROJECT_NAME})
 add_executable(speakers_node src/speakers_node.cpp)
 ament_target_dependencies(speakers_node rclcpp audio_msgs)
 target_link_libraries(speakers_node ${PROJECT_NAME})
-
-
-
-
 
 add_executable(test_motors test/test_motors.cpp src/rae_motors.cpp)
 target_link_libraries(test_motors ${GPIOD_LIBRARY})
@@ -113,8 +100,6 @@ install(
   DIRECTORY include/
   DESTINATION include
 )
-
-
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
