+++ rae_hw/include/rae_hw/peripherals/led.hpp
@@ -14,6 +14,7 @@
 #include <linux/ioctl.h>
 #include <sys/stat.h>
 #include <map>
+#include <cmath>
 #include <linux/types.h>
 #include "rae_hw/peripherals/spidev.h"
 #include "rclcpp/rclcpp.hpp"
@@ -36,13 +37,13 @@ namespace rae_hw
     private:
         std::map<uint16_t, uint16_t> logicalToPhysicalMapping;
         void transmitSPI();
-        void fillBuffer(uint8_t color);
+        void fillBuffer(uint8_t color, float intensity = 1);
         void LED_control ();
         uint8_t convertColor(float num);
-        void setSinglePixel(uint16_t pixel, uint8_t r, uint8_t g, uint8_t b, float period = 0.0);
-        void setAllPixels(uint8_t r, uint8_t g, uint8_t b, float period = 0.0);
-        void spinner(uint8_t r, uint8_t g, uint8_t b,  uint8_t size = 5 , uint8_t blades = 1, float period = 0.0);
-        void fan(uint8_t r, uint8_t g, uint8_t b,  bool opening, uint8_t blades = 1, float period = 0.0);
+        void setSinglePixel(uint16_t pixel, uint8_t r, uint8_t g, uint8_t b, float frequency = 0.0);
+        void setAllPixels(uint8_t r, uint8_t g, uint8_t b, float frequency = 0.0);
+        void spinner(uint8_t r, uint8_t g, uint8_t b,  uint8_t size = 5 , uint8_t blades = 1, float frequency = 0.0);
+        void fan(uint8_t r, uint8_t g, uint8_t b,  bool opening, uint8_t blades = 1, float frequency = 0.0);
         void topic_callback(const rae_msgs::msg::LEDControl::SharedPtr msg); 
 
         rclcpp::Subscription<rae_msgs::msg::LEDControl>::SharedPtr subscription_;
