+++ rae_msgs/srv/RecordAudio.srv
@@ -1,0 +1,3 @@
+string file_location
+---
+bool success
